% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_analysis.R
\name{boxplots_by_clust}
\alias{boxplots_by_clust}
\title{Plot cluster variables}
\usage{
boxplots_by_clust(data_col, cluster_col)
}
\arguments{
\item{cluster_col}{}
}
\description{
Plot cluster variables
}
