% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_analysis.R
\name{compute_kclust}
\alias{compute_kclust}
\title{Compute k-means clusters}
\usage{
compute_kclust(input, cluster_number)
}
\arguments{
\item{cluster_number}{}
}
\description{
Compute k-means clusters
}
