% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_relative.R
\name{compute_relative}
\alias{compute_relative}
\title{Compute relative values}
\usage{
compute_relative(input)
}
\arguments{
\item{input}{An R vector.}
}
\value{
Returns a vector with relative values based on maximal values from
the given vector.
}
\description{
Compute relative values
}
