% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBM_workflow.R
\name{gbm_data_partition}
\alias{gbm_data_partition}
\title{Creates train and test data sets}
\usage{
gbm_data_partition(input, sep_col, sep_prop)
}
\arguments{
\item{sep_prop}{}
}
\description{
Creates train and test data sets
}
