% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBM_workflow.R
\name{lgbm_export}
\alias{lgbm_export}
\title{Export the best LGBM parameters}
\usage{
lgbm_export(study, my_date, name_seq, lgbm_model_results)
}
\arguments{
\item{lgbm_model_results}{}
}
\description{
Export the best LGBM parameters
}
