% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBM_workflow.R
\name{lgbm_round_tune}
\alias{lgbm_round_tune}
\title{Compute the optimal rounds for a LGBM prediction}
\usage{
lgbm_round_tune(
  my_rounds,
  lgbm_params,
  lgbm_dtrain,
  lgbm_valids,
  lgbm_test_data,
  lgbm_test_data_label
)
}
\arguments{
\item{lgbm_test_data_label}{}
}
\description{
Compute the optimal rounds for a LGBM prediction
}
