% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_analysis.R
\name{optimal_clust}
\alias{optimal_clust}
\title{Compute optimal cluster numbers}
\usage{
optimal_clust(input_data, cluster_method)
}
\arguments{
\item{cluster_method}{}
}
\description{
Compute optimal cluster numbers
}
