% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_analysis.R
\name{subject_repartition}
\alias{subject_repartition}
\title{Plot subject repartition in clusters}
\usage{
subject_repartition(data_col, cluster_col)
}
\arguments{
\item{cluster_col}{}
}
\description{
Plot subject repartition in clusters
}
